#46: Refactored packages around for settings ui
